Remove dead api token
Shouldn't have checked this in in the first place, but it's a
client-side token so no issue with it being public anyway. I have
since deleted the Google project that had the key, so this key should
be dead and there is no need to go back in version history to strip
it from past revisions.